requirements.txt
aiohttp>=3.10,<4
discord.py>=2.4,<3
python-dotenv>=1.0,<2
requests>=2.31,<3
